what if i stopped maintaining the other themes...
